Add files via upload
The project aims to create a user-friendly Python-based application with a simple and intuitive user interface to allow users to add and view expenses, including the date and expense amount. It will enable users to view their expense history in a structured format and store expense data in an Excel spreadsheet for easy access and management, simplifying the process of tracking daily expenses for  individuals seeking a straightforward financial management solution.